+++ previous-versions/files/pkg/models/shared/file.go
@@ -2,10 +2,18 @@
 
 package shared
 
+import (
+	"github.com/codatio/client-sdk-go/previous-versions/files/pkg/utils"
+)
+
 type File struct {
+	AdditionalProperties map[string]interface{} `additionalProperties:"true" json:"-"`
+	// An optional display name for the file.
 	DisplayName *string `json:"displayName,omitempty"`
-	FileName    *string `json:"fileName,omitempty"`
-	SourceType  *string `json:"sourceType,omitempty"`
+	// The file's name.
+	FileName *string `json:"fileName,omitempty"`
+	// The source of the file uploaded.
+	SourceType *string `json:"sourceType,omitempty"`
 	// In Codat's data model, dates and times are represented using the <a class="external" href="https://en.wikipedia.org/wiki/ISO_8601" target="_blank">ISO 8601 standard</a>. Date and time fields are formatted as strings; for example:
 	//
 	// ```
@@ -26,6 +34,24 @@ type File struct {
 	// > Not all dates from Codat will contain information about time zones.
 	// > Where it is not available from the underlying platform, Codat will return these as times local to the business whose data has been synced.
 	Uploaded *string `json:"uploaded,omitempty"`
+}
+
+func (f File) MarshalJSON() ([]byte, error) {
+	return utils.MarshalJSON(f, "", false)
+}
+
+func (f *File) UnmarshalJSON(data []byte) error {
+	if err := utils.UnmarshalJSON(data, &f, "", false, false); err != nil {
+		return err
+	}
+	return nil
+}
+
+func (o *File) GetAdditionalProperties() map[string]interface{} {
+	if o == nil {
+		return nil
+	}
+	return o.AdditionalProperties
 }
 
 func (o *File) GetDisplayName() *string {