+++ previous-versions/files/RELEASES.md
@@ -299,3 +299,13 @@ Based on:
 - [go v0.23.0] previous-versions/files
 ### Releases
 - [Go v0.23.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/files/v0.23.0 - previous-versions/files
+
+## 2023-10-09 01:24:50
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Files.yaml
+- Speakeasy CLI 1.96.1 (2.150.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [go v0.24.0] previous-versions/files
+### Releases
+- [Go v0.24.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/files/v0.24.0 - previous-versions/files